requirements.txt
torch>=2.0.0
diffusers>=0.20.0
transformers>=4.25.0
accelerate>=0.16.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
controlnet-aux>=0.0.7
